requirements.txt
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
pandas>=2.1.0
numpy>=1.26.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
dash>=2.17.0
dash-bootstrap-components>=1.6.0
plotly>=5.22.0
groq>=0.9.0
python-docx>=1.1.0
requests>=2.31.0
gunicorn>=22.0.0
